+++ test/test_stl_triangle_binary.cpp
@@ -9,16 +9,16 @@ using namespace std;
 int main(void)
 {
 
-	const double n1[] = { 0.0, 0.0, -1.0 };
-	const double v1[] = {
+	const float n1[] = { 0.0, 0.0, -1.0 };
+	const float v1[] = {
 		3.10000e+1, 4.15500e+1, 1.00000e+0,
 		3.10000e+1, 1.00000e+1, 1.00000e+0,
 		6.00000e+0, 2.50000e-1, 1.00000e+0
 	};
 	STL_triangle tri1(v1, n1);
 
-	const double n2[] = { 0.0, 0.0, -1.0 };
-	const double v2[] = {
+	const float n2[] = { 0.0, 0.0, -1.0 };
+	const float v2[] = {
 		3.10000e+001, 4.15500e+001, 1.00000e+000,
 		6.10000e+001, 2.50000e-001, 1.00000e+000,
 		3.10000e+001, 1.00000e+001, 1.00000e+000
